@@ -2,22 +2,24 @@
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
     <TargetLatestRuntimePatch>true</TargetLatestRuntimePatch>
-    <AssemblyVersion>1.5.3.0</AssemblyVersion>
-    <Version>1.5.3.0</Version>
-    <FileVersion>1.5.3.0</FileVersion>
-    <PackageVersion>1.5.3.0</PackageVersion>
+    <AssemblyVersion>1.6.0.0</AssemblyVersion>
+    <Version>1.6.9.0</Version>
+    <FileVersion>1.6.0.0</FileVersion>
+    <PackageVersion>1.6.0.0</PackageVersion>
+    <LangVersion>11</LangVersion>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
     <None Remove=".DS_Store" />
     <None Remove="System.Data.SqlClient" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
     <PackageReference Include="Newtonsoft.Json.Bson" Version="1.0.2" />
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
-    <PackageReference Include="System.Data.SqlClient" Version="4.8.3" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="7.0.0" />
+    <PackageReference Include="System.Data.SqlClient" Version="4.8.5" />
   </ItemGroup>
 </Project>
