@@ -4,186 +4,167 @@
 using System;
 using System.IO;
 using System.Threading;
-using Fynydd.Halide;
+using SQribe.Halide.Core;
 
-namespace SQribe
+namespace SQribe;
+
+public interface IFullTextCatalogs
 {
-    public interface IFullTextCatalogs
-	{
-        ISettings settings { get; }
+    ISettings settings { get; }
 
-        IOutput output { get; }
+    IOutput output { get; }
 
-        IHelpers helpers { get; }
+    IHelpers helpers { get; }
 
-        void GenerateCreateScript(ref int counter, ref int total, long groupToken);
+    void GenerateCreateScript(ref int counter, ref int total, long groupToken);
 
-        void DropAll(long token);
+    void DropAll(long token);
 
-        void Restore(long token);
+    void Restore(long token);
+}
+
+
+public class FullTextCatalogs : IFullTextCatalogs
+{
+    #region Public properties
+
+    public ISettings settings => _settings;
+
+    public IOutput output => _output;
+
+    public IHelpers helpers => _helpers;
+
+    #endregion
+
+    #region Private properties
+
+    private readonly ISettings _settings;
+    private readonly IOutput _output;
+    private readonly IHelpers _helpers;
+
+    #endregion
+
+    public FullTextCatalogs(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
+    {
+        _settings = singletonSettings;
+        _output = singletonOutput;
+        _helpers = singletonHelpers;
     }
 
-
-	public class FullTextCatalogs : IFullTextCatalogs
-	{
-        #region Public properties
-
-        public ISettings settings
+    /// <summary>
+    /// Generate script to create fulltext catalogs and indexes functions.
+    /// </summary>
+    public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
+    {
+        if (settings.SqlObjects.Contains(",ftc,") && settings.Abort == false)
         {
-            get
-            {
-                return _settings;
-            }
-        }
+            const string objectName = "fulltext catalog";
+            var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
+            var startDate = DateTime.Now;
+            var lastTimeUpdate = string.Empty;
+            var currentCount = 0;
+            var totalCount = 0;
 
-        public IOutput output
-        {
-            get
-            {
-                return _output;
-            }
-        }
-
-        public IHelpers helpers
-        {
-            get
-            {
-                return _helpers;
-            }
-        }
-
-        #endregion
-
-        #region Private properties
-
-        private readonly ISettings _settings;
-        private readonly IOutput _output;
-        private readonly IHelpers _helpers;
-
-        #endregion
-
-        public FullTextCatalogs(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
-        {
-            _settings = singletonSettings;
-            _output = singletonOutput;
-            _helpers = singletonHelpers;
-        }
-
-		/// <summary>
-		/// Generate script to create fulltext catalogs and indexes functions.
-		/// </summary>
-        public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
-        {
-            if (settings.SqlObjects.Contains(",ftc,") && settings.Abort == false)
-            {
-                var objectName = "fulltext catalog";
-                var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
-                var startDate = DateTime.Now;
-                var lastTimeUpdate = "";
-                int currentCount = 0;
-                int totalCount = 0;
-
-                helpers.GenerateCreateScript (
-                    objectName, 
-                    settings.OutputPath + settings.FulltextCatalogsFilename, 
-                    (script, token) => {
+            helpers.GenerateCreateScript (
+                objectName, 
+                settings.OutputPath + settings.FulltextCatalogsFilename, 
+                (script, token) => {
                         
-                        using (var reader = new DataReader(helpers.LoadScript("generate-fulltext-catalogs.sql"), settings.DataSource, useRewind: true))
+                    using (var reader = new DataReader(helpers.LoadScript("generate-fulltext-catalogs.sql"), settings.DataSource, useRewind: true))
+                    {
+                        if (settings.Abort == false)
                         {
-                            if (settings.Abort == false)
+                            var cts = new CancellationTokenSource();
+                            var task = reader.ExecuteAsync(cts.Token);
+
+                            while (task.IsCompleted == false)
                             {
-                                var cts = new CancellationTokenSource();
-                                var task = reader.ExecuteAsync(cts.Token);
-
-                                while (task.IsCompleted == false)
+                                if (settings.Abort)
                                 {
-                                    if (settings.Abort)
-                                    {
-                                        cts.Cancel();
-                                    }
-
-                                    System.Threading.Thread.Sleep(Constants.SleepNumber);
+                                    cts.Cancel();
                                 }
 
-                                if (reader.IsReady)
-                                {
-                                    if (reader.HasRows)
-                                    {
-                                        do
-                                        {
-                                            while (reader.Read() && settings.Abort == false)
-                                            {
-                                                totalCount++;
-                                            }
-
-                                            System.Threading.Thread.Sleep(Constants.SleepNumber);
-
-                                        } while (reader.NextResult() && settings.Abort == false);
-                                    }
-                                }
+                                Thread.Sleep(Constants.SleepNumber);
                             }
 
-                            if (settings.Abort == false)
+                            if (reader.IsReady)
                             {
-                                if (reader.Rewind())
+                                if (reader.HasRows)
                                 {
-                                    if (reader.HasRows)
+                                    do
                                     {
-                                        do
+                                        while (reader.Read() && settings.Abort == false)
                                         {
-                                            while (reader.Read() && settings.Abort == false)
-                                            {
-                                                if (reader[0].StartsWith("CREATE FULLTEXT CATALOG"))
-                                                {
-                                                    currentCount++;
+                                            totalCount++;
+                                        }
 
-                                                    script += "-- SQRIBE/OBJ;" + settings.Hash + Constants.LineFeed;
-                                                }
+                                        Thread.Sleep(Constants.SleepNumber);
 
-                                                script += reader[0] + Constants.LineFeed;
-
-                                                helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
-                                            }
-
-                                            System.Threading.Thread.Sleep(Constants.SleepNumber);
-
-                                        } while (reader.NextResult() && settings.Abort == false);
-                                    }
+                                    } while (reader.NextResult() && settings.Abort == false);
                                 }
                             }
                         }
 
-                        if (currentCount > 0)
+                        if (settings.Abort == false)
                         {
-                            script = script.StandardizeLineEndings().Replace("GO" + Constants.LineFeed, "GO -- SQRIBE/GO;" + settings.Hash + Constants.LineFeed);
+                            if (reader.Rewind())
+                            {
+                                if (reader.HasRows)
+                                {
+                                    do
+                                    {
+                                        while (reader.Read() && settings.Abort == false)
+                                        {
+                                            if (reader[0].StartsWith("CREATE FULLTEXT CATALOG"))
+                                            {
+                                                currentCount++;
+
+                                                script += "-- SQRIBE/OBJ;" + settings.Hash + Constants.LineFeed;
+                                            }
+
+                                            script += reader[0] + Constants.LineFeed;
+
+                                            helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
+                                        }
+
+                                        Thread.Sleep(Constants.SleepNumber);
+
+                                    } while (reader.NextResult() && settings.Abort == false);
+                                }
+                            }
                         }
+                    }
 
-                        return Tuple.Create(script, currentCount);
-                    },
-                    ref counter, ref total, groupToken
-                );
-            }
-        }
+                    if (currentCount > 0)
+                    {
+                        script = script.StandardizeLineEndings().Replace("GO" + Constants.LineFeed, "GO -- SQRIBE/GO;" + settings.Hash + Constants.LineFeed);
+                    }
 
-        public void DropAll(long token)
-        {
-            helpers.DropObject(
-                token, 
-                ",ftc,", 
-                settings.ScriptPath + settings.FulltextCatalogsFilename, 
-                "fulltext catalog", 
-                "drops" + Path.DirectorySeparatorChar.ToString() + "drop-fulltext-catalogs.sql"
-                );
+                    return Tuple.Create(script, currentCount);
+                },
+                ref counter, ref total, groupToken
+            );
         }
+    }
 
-        public void Restore(long token)
-        {
-            helpers.RestoreObject(
-                token, 
-                ",ftc,", 
-                "fulltext catalog", 
-                settings.ScriptPath + settings.FulltextCatalogsFilename
-                );
-        }
+    public void DropAll(long token)
+    {
+        helpers.DropObject(
+            token, 
+            ",ftc,", 
+            settings.ScriptPath + settings.FulltextCatalogsFilename, 
+            "fulltext catalog", 
+            "drops" + Path.DirectorySeparatorChar + "drop-fulltext-catalogs.sql"
+        );
+    }
+
+    public void Restore(long token)
+    {
+        helpers.RestoreObject(
+            token, 
+            ",ftc,", 
+            "fulltext catalog", 
+            settings.ScriptPath + settings.FulltextCatalogsFilename
+        );
     }
 }
