@@ -4,168 +4,150 @@
 using System;
 using System.IO;
 using System.Threading;
-using Fynydd.Halide;
+using SQribe.Halide.Core;
 
-namespace SQribe
+namespace SQribe;
+
+public interface IStoredProcedures
 {
-    public interface IStoredProcedures
-	{
-        ISettings settings { get; }
+    ISettings settings { get; }
 
-        IOutput output { get; }
+    IOutput output { get; }
 
-        IHelpers helpers { get; }
+    IHelpers helpers { get; }
 
-        void GenerateCreateScript(ref int counter, ref int total, long groupToken);
+    void GenerateCreateScript(ref int counter, ref int total, long groupToken);
 
-        void DropAll(long token);
+    void DropAll(long token);
 
-        void Restore(long token);
+    void Restore(long token);
+}
+
+
+public class StoredProcedures : IStoredProcedures
+{
+    #region Public properties
+
+    public ISettings settings => _settings;
+
+    public IOutput output => _output;
+
+    public IHelpers helpers => _helpers;
+
+    #endregion
+
+    #region Private properties
+
+    private readonly ISettings _settings;
+    private readonly IOutput _output;
+    private readonly IHelpers _helpers;
+
+    #endregion
+
+    public StoredProcedures(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
+    {
+        _settings = singletonSettings;
+        _output = singletonOutput;
+        _helpers = singletonHelpers;
     }
 
-
-	public class StoredProcedures : IStoredProcedures
-	{
-        #region Public properties
-
-        public ISettings settings
+    /// <summary>
+    /// Generate script to create stored procedures.
+    /// </summary>
+    public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
+    {
+        if (settings.SqlObjects.Contains(",usp,") && settings.Abort == false)
         {
-            get
-            {
-                return _settings;
-            }
-        }
+            const string objectName = "stored procedure";
+            var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
+            var startDate = DateTime.Now;
+            var lastTimeUpdate = string.Empty;
+            var currentCount = 0;
+            var totalCount = 0;
 
-        public IOutput output
-        {
-            get
-            {
-                return _output;
-            }
-        }
+            helpers.GenerateCreateScript (
+                objectName, 
+                settings.OutputPath + settings.StoredProceduresFilename, 
+                (script, token) => {
 
-        public IHelpers helpers
-        {
-            get
-            {
-                return _helpers;
-            }
-        }
+                    using (var reader = new DataReader(helpers.LoadScript("generate-stored-procedures.sql"), settings.DataSource, useRewind: true))
+                    {
+                        if (settings.Abort == false)
+                        {
+                            var cts = new CancellationTokenSource();
+                            var task = reader.ExecuteAsync(cts.Token);
 
-        #endregion
-
-        #region Private properties
-
-        private readonly ISettings _settings;
-        private readonly IOutput _output;
-        private readonly IHelpers _helpers;
-
-        #endregion
-
-        public StoredProcedures(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
-        {
-            _settings = singletonSettings;
-            _output = singletonOutput;
-            _helpers = singletonHelpers;
-        }
-
-		/// <summary>
-		/// Generate script to create stored procedures.
-		/// </summary>
-        public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
-        {
-            if (settings.SqlObjects.Contains(",usp,") && settings.Abort == false)
-            {
-                var objectName = "stored procedure";
-                var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
-                var startDate = DateTime.Now;
-                var lastTimeUpdate = "";
-                int currentCount = 0;
-                int totalCount = 0;
-
-                helpers.GenerateCreateScript (
-                    objectName, 
-                    settings.OutputPath + settings.StoredProceduresFilename, 
-                    (script, token) => {
-
-                            using (var reader = new DataReader(helpers.LoadScript("generate-stored-procedures.sql"), settings.DataSource, useRewind: true))
+                            while (task.IsCompleted == false)
                             {
-                                if (settings.Abort == false)
+                                if (settings.Abort)
                                 {
-                                    var cts = new CancellationTokenSource();
-                                    var task = reader.ExecuteAsync(cts.Token);
-
-                                    while (task.IsCompleted == false)
-                                    {
-                                        if (settings.Abort)
-                                        {
-                                            cts.Cancel();
-                                        }
-
-                                        System.Threading.Thread.Sleep(Constants.SleepNumber);
-                                    }
-
-                                    if (reader.IsReady)
-                                    {
-                                        if (reader.HasRows)
-                                        {
-                                            while (reader.Read() && settings.Abort == false)
-                                            {
-                                                totalCount++;
-                                            }
-                                        }
-                                    }
+                                    cts.Cancel();
                                 }
 
-                                if (settings.Abort == false)
+                                Thread.Sleep(Constants.SleepNumber);
+                            }
+
+                            if (reader.IsReady)
+                            {
+                                if (reader.HasRows)
                                 {
-                                    if (reader.Rewind())
+                                    while (reader.Read() && settings.Abort == false)
                                     {
-                                        if (reader.HasRows)
-                                        {
-                                            while (reader.Read() && settings.Abort == false)
-                                            {
-                                                var val = reader[0].ToString();
-                                                script += val
-                                                            .Replace(Constants.LineFeed + Constants.LineFeed + "CREATE PROCEDURE", Constants.LineFeed + "CREATE PROCEDURE")
-                                                            .Replace(Constants.LineFeed + Constants.LineFeed + "END" + Constants.LineFeed, Constants.LineFeed + "END" + Constants.LineFeed)
-                                                            .Replace(Constants.LineFeed + Constants.LineFeed + "GO -- SQRIBE/GO", Constants.LineFeed + "GO -- SQRIBE/GO");
-
-                                                currentCount++;
-
-                                                helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
-                                            }
-                                        }
+                                        totalCount++;
                                     }
                                 }
                             }
+                        }
+
+                        if (settings.Abort == false)
+                        {
+                            if (reader.Rewind())
+                            {
+                                if (reader.HasRows)
+                                {
+                                    while (reader.Read() && settings.Abort == false)
+                                    {
+                                        var val = reader[0];
+                                        
+                                        script += val
+                                            .Replace(Constants.LineFeed + Constants.LineFeed + "CREATE PROCEDURE", Constants.LineFeed + "CREATE PROCEDURE")
+                                            .Replace(Constants.LineFeed + Constants.LineFeed + "END" + Constants.LineFeed, Constants.LineFeed + "END" + Constants.LineFeed)
+                                            .Replace(Constants.LineFeed + Constants.LineFeed + "GO -- SQRIBE/GO", Constants.LineFeed + "GO -- SQRIBE/GO");
+
+                                        currentCount++;
+
+                                        helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
+                                    }
+                                }
+                            }
+                        }
+                    }
                         
-                            return Tuple.Create(script, currentCount);
-                        },
-                        ref counter, ref total, groupToken
-                    );
-            }
+                    return Tuple.Create(script, currentCount);
+                },
+                ref counter, ref total, groupToken
+            );
         }
+    }
 
-        public void DropAll(long token)
-        {
-            helpers.DropObject(
-                token, 
-                ",usp,", 
-                settings.ScriptPath + settings.StoredProceduresFilename, 
-                "stored procedure", 
-                "drops" + Path.DirectorySeparatorChar.ToString() + "drop-stored-procedures.sql"
-                );
-        }
+    public void DropAll(long token)
+    {
+        helpers.DropObject(
+            token, 
+            ",usp,", 
+            settings.ScriptPath + settings.StoredProceduresFilename, 
+            "stored procedure", 
+            "drops" + Path.DirectorySeparatorChar + "drop-stored-procedures.sql"
+        );
+    }
 
-        public void Restore(long token)
-        {
-            helpers.RestoreObject(
-                token, 
-                ",usp,", 
-                "stored procedure", 
-                settings.ScriptPath + settings.StoredProceduresFilename
-                );
-        }
+    public void Restore(long token)
+    {
+        helpers.RestoreObject(
+            token, 
+            ",usp,", 
+            "stored procedure", 
+            settings.ScriptPath + settings.StoredProceduresFilename
+        );
     }
 }
