@@ -5,183 +5,166 @@ using System;
 using System.IO;
 using System.Text.RegularExpressions;
 using System.Threading;
-using Fynydd.Halide;
+using SQribe.Halide.Core;
 
-namespace SQribe
+namespace SQribe;
+
+public interface IViews
 {
-    public interface IViews
-	{
-        ISettings settings { get; }
+    ISettings settings { get; }
 
-        IOutput output { get; }
+    IOutput output { get; }
 
-        IHelpers helpers { get; }
+    IHelpers helpers { get; }
 
-        void GenerateCreateScript(ref int counter, ref int total, long groupToken);
+    void GenerateCreateScript(ref int counter, ref int total, long groupToken);
 
-        void DropAll(long token);
+    void DropAll(long token);
 
-        void Restore(long token);
+    void Restore(long token);
+}
+
+public class Views : IViews
+{
+    #region Public properties
+
+    public ISettings settings => _settings;
+
+    public IOutput output => _output;
+
+    public IHelpers helpers => _helpers;
+
+    #endregion
+
+    #region Private properties
+
+    private readonly ISettings _settings;
+    private readonly IOutput _output;
+    private readonly IHelpers _helpers;
+
+    #endregion
+
+    public Views(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
+    {
+        _settings = singletonSettings;
+        _output = singletonOutput;
+        _helpers = singletonHelpers;
     }
 
-	public class Views : IViews
-	{
-        #region Public properties
-
-        public ISettings settings
+    /// <summary>
+    /// Generate script to create views.
+    /// </summary>
+    public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
+    {
+        if (settings.SqlObjects.Contains(",view,") && settings.Abort == false)
         {
-            get
-            {
-                return _settings;
-            }
-        }
+            const string objectName = "view";
+            var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
+            var startDate = DateTime.Now;
+            var lastTimeUpdate = string.Empty;
+            var currentCount = 0;
+            var totalCount = 0;
 
-        public IOutput output
-        {
-            get
-            {
-                return _output;
-            }
-        }
+            helpers.GenerateCreateScript (
+                objectName, 
+                settings.OutputPath + settings.ViewsFilename, 
+                (script, token) => {
 
-        public IHelpers helpers
-        {
-            get
-            {
-                return _helpers;
-            }
-        }
-
-        #endregion
-
-        #region Private properties
-
-        private readonly ISettings _settings;
-        private readonly IOutput _output;
-        private readonly IHelpers _helpers;
-
-        #endregion
-
-        public Views(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
-        {
-            _settings = singletonSettings;
-            _output = singletonOutput;
-            _helpers = singletonHelpers;
-        }
-
-		/// <summary>
-		/// Generate script to create views.
-		/// </summary>
-        public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
-        {
-            if (settings.SqlObjects.Contains(",view,") && settings.Abort == false)
-            {
-                var objectName = "view";
-                var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
-                var startDate = DateTime.Now;
-                var lastTimeUpdate = "";
-                int currentCount = 0;
-                int totalCount = 0;
-
-                helpers.GenerateCreateScript (
-                    objectName, 
-                    settings.OutputPath + settings.ViewsFilename, 
-                    (script, token) => {
-
-                        using (var reader = new DataReader(helpers.LoadScript("generate-views.sql"), settings.DataSource, useRewind: true))
+                    using (var reader = new DataReader(helpers.LoadScript("generate-views.sql"), settings.DataSource, useRewind: true))
+                    {
+                        if (settings.Abort == false)
                         {
-                            if (settings.Abort == false)
+                            var cts = new CancellationTokenSource();
+                            var task = reader.ExecuteAsync(cts.Token);
+
+                            while (task.IsCompleted == false)
                             {
-                                var cts = new CancellationTokenSource();
-                                var task = reader.ExecuteAsync(cts.Token);
-
-                                while (task.IsCompleted == false)
+                                if (settings.Abort)
                                 {
-                                    if (settings.Abort)
-                                    {
-                                        cts.Cancel();
-                                    }
-
-                                    System.Threading.Thread.Sleep(Constants.SleepNumber);
+                                    cts.Cancel();
                                 }
 
-                                if (reader.IsReady)
-                                {
-                                    if (reader.HasRows)
-                                    {
-                                        while(reader.Read() && settings.Abort == false)
-                                        {
-                                            totalCount++;
-                                        }
-                                    }
-                                }
+                                Thread.Sleep(Constants.SleepNumber);
                             }
 
-                            if (settings.Abort == false)
+                            if (reader.IsReady)
                             {
-                                if (reader.Rewind())
+                                if (reader.HasRows)
                                 {
-                                    if (reader.HasRows)
+                                    while(reader.Read() && settings.Abort == false)
                                     {
-                                        while(reader.Read() && settings.Abort == false)
-                                        {
-                                            var val = reader[0].ToString();
-
-                                            #region Look for errant XML schema paths and fix
-
-                                            // Prevents XML singleton errors
-                                            if (val.Contains("].ref.value(N'"))
-                                            {
-                                                string pattern = @"\]\.ref\.value\(N'.*?(?<!\])',";
-
-                                                Regex regEx = new Regex(pattern, RegexOptions.IgnoreCase | RegexOptions.Singleline | RegexOptions.IgnorePatternWhitespace);
-                                                MatchCollection matches = regEx.Matches(val);
-
-                                                foreach (var match in matches)
-                                                {
-                                                    val = val.Replace(match.ToString(), match.ToString().Left(match.ToString().Length - 2) + "[1]',");
-                                                }
-                                            }
-
-                                            #endregion
-
-                                            script += val;
-
-                                            currentCount++;
-
-                                            helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
-                                        }
+                                        totalCount++;
                                     }
                                 }
                             }
                         }
+
+                        if (settings.Abort == false)
+                        {
+                            if (reader.Rewind())
+                            {
+                                if (reader.HasRows)
+                                {
+                                    while(reader.Read() && settings.Abort == false)
+                                    {
+                                        var val = reader[0];
+
+                                        #region Look for errant XML schema paths and fix
+
+                                        // Prevents XML singleton errors
+                                        if (val.Contains("].ref.value(N'"))
+                                        {
+                                            var pattern = @"\]\.ref\.value\(N'.*?(?<!\])',";
+                                            var regEx = new Regex(pattern, RegexOptions.IgnoreCase | RegexOptions.Singleline | RegexOptions.IgnorePatternWhitespace);
+                                            var matches = regEx.Matches(val);
+
+                                            foreach (var match in matches)
+                                            {
+                                                var matchStr = match.ToString();
+                                                
+                                                if (matchStr != null)
+                                                    val = val.Replace(matchStr, matchStr.Left(matchStr.Length - 2) + "[1]',");
+                                            }
+                                        }
+
+                                        #endregion
+
+                                        script += val;
+
+                                        currentCount++;
+
+                                        helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
+                                    }
+                                }
+                            }
+                        }
+                    }
                         
-                        return Tuple.Create(script, currentCount);
-                    },
-                    ref counter, ref total, groupToken
-                );
-            }
+                    return Tuple.Create(script, currentCount);
+                },
+                ref counter, ref total, groupToken
+            );
         }
+    }
 
-        public void DropAll(long token)
-        {
-            helpers.DropObject(
-                token, 
-                ",view,", 
-                settings.ScriptPath + settings.ViewsFilename, 
-                "view", 
-                "drops" + Path.DirectorySeparatorChar.ToString() + "drop-views.sql"
-                );
-        }
+    public void DropAll(long token)
+    {
+        helpers.DropObject(
+            token, 
+            ",view,", 
+            settings.ScriptPath + settings.ViewsFilename, 
+            "view", 
+            "drops" + Path.DirectorySeparatorChar + "drop-views.sql"
+        );
+    }
 
-        public void Restore(long token)
-        {
-            helpers.RestoreObject(
-                token, 
-                ",view,", 
-                "view", 
-                settings.ScriptPath + settings.ViewsFilename
-                );
-        }
+    public void Restore(long token)
+    {
+        helpers.RestoreObject(
+            token, 
+            ",view,", 
+            "view", 
+            settings.ScriptPath + settings.ViewsFilename
+        );
     }
 }
