@@ -4,169 +4,149 @@
 using System;
 using System.IO;
 using System.Threading;
-using Fynydd.Halide;
+using SQribe.Halide.Core;
 
-namespace SQribe
+namespace SQribe;
+
+public interface IXmlSchemaCollections
 {
-    public interface IXmlSchemaCollections
-	{
-        ISettings settings { get; }
+    ISettings settings { get; }
 
-        IOutput output { get; }
+    IOutput output { get; }
 
-        IHelpers helpers { get; }
+    IHelpers helpers { get; }
 
-        void GenerateCreateScript(ref int counter, ref int total, long groupToken);
+    void GenerateCreateScript(ref int counter, ref int total, long groupToken);
 
-        void DropAll(long token);
+    void DropAll(long token);
 
-        void Restore(long token);
+    void Restore(long token);
+}
+
+public class XmlSchemaCollections : IXmlSchemaCollections
+{
+    #region Public properties
+
+    public ISettings settings => _settings;
+
+    public IOutput output => _output;
+
+    public IHelpers helpers => _helpers;
+
+    #endregion
+
+    #region Private properties
+
+    private readonly ISettings _settings;
+    private readonly IOutput _output;
+    private readonly IHelpers _helpers;
+
+    #endregion
+
+    public XmlSchemaCollections(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
+    {
+        _settings = singletonSettings;
+        _output = singletonOutput;
+        _helpers = singletonHelpers;
     }
 
-	public class XmlSchemaCollections : IXmlSchemaCollections
-	{
-        #region Public properties
-
-        public ISettings settings
+    /// <summary>
+    /// Generate script to create XML schema collections.
+    /// </summary>
+    public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
+    {
+        if (settings.SqlObjects.Contains(",xsc,") && settings.Abort == false)
         {
-            get
-            {
-                return _settings;
-            }
-        }
+            const string objectName = "XML schema collection";
+            var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
+            var startDate = DateTime.Now;
+            var lastTimeUpdate = string.Empty;
+            var currentCount = 0;
+            var totalCount = 0;
 
-        public IOutput output
-        {
-            get
-            {
-                return _output;
-            }
-        }
-
-        public IHelpers helpers
-        {
-            get
-            {
-                return _helpers;
-            }
-        }
-
-        #endregion
-
-        #region Private properties
-
-        private readonly ISettings _settings;
-        private readonly IOutput _output;
-        private readonly IHelpers _helpers;
-
-        #endregion
-
-        public XmlSchemaCollections(ISettings singletonSettings, IOutput singletonOutput, IHelpers singletonHelpers)
-        {
-            _settings = singletonSettings;
-            _output = singletonOutput;
-            _helpers = singletonHelpers;
-        }
-
-		/// <summary>
-		/// Generate script to create XML schema collections.
-		/// </summary>
-        public void GenerateCreateScript(ref int counter, ref int total, long groupToken)
-        {
-            if (settings.SqlObjects.Contains(",xsc,") && settings.Abort == false)
-            {
-                var objectName = "XML schema collection";
-                var prefix = objectName.PluralizeNoun(2).ToUpperFirstCharacter();
-                var startDate = DateTime.Now;
-                var lastTimeUpdate = "";
-                int currentCount = 0;
-                int totalCount = 0;
-
-                helpers.GenerateCreateScript (
-                    objectName, 
-                    settings.OutputPath + settings.XmlSchemaCollectionsFilename, 
-                    (script, token) => {
+            helpers.GenerateCreateScript (
+                objectName, 
+                settings.OutputPath + settings.XmlSchemaCollectionsFilename, 
+                (script, token) => {
                         
-                        var template = helpers.LoadTemplate("create-xml-schema-collection.sql");
+                    var template = helpers.LoadTemplate("create-xml-schema-collection.sql");
 
-                        using (var reader = new DataReader(helpers.LoadScript("select-xml-schema-collections.sql"), settings.DataSource, useRewind: true))
+                    using (var reader = new DataReader(helpers.LoadScript("select-xml-schema-collections.sql"), settings.DataSource, useRewind: true))
+                    {
+                        if (settings.Abort == false)
                         {
-                            if (settings.Abort == false)
+                            var cts = new CancellationTokenSource();
+                            var task = reader.ExecuteAsync(cts.Token);
+
+                            while (task.IsCompleted == false)
                             {
-                                var cts = new CancellationTokenSource();
-                                var task = reader.ExecuteAsync(cts.Token);
-
-                                while (task.IsCompleted == false)
+                                if (settings.Abort)
                                 {
-                                    if (settings.Abort)
-                                    {
-                                        cts.Cancel();
-                                    }
-
-                                    System.Threading.Thread.Sleep(Constants.SleepNumber);
+                                    cts.Cancel();
                                 }
 
-                                if (reader.IsReady)
-                                {
-                                    if (reader.HasRows)
-                                    {
-                                        while(reader.Read() && settings.Abort == false)
-                                        {
-                                            totalCount++;
-                                        }
-                                    }
-                                }
+                                Thread.Sleep(Constants.SleepNumber);
                             }
 
-                            if (settings.Abort == false)
+                            if (reader.IsReady)
                             {
-                                if (reader.Rewind())
+                                if (reader.HasRows)
                                 {
-                                    if (reader.HasRows)
+                                    while(reader.Read() && settings.Abort == false)
                                     {
-                                        while(reader.Read() && settings.Abort == false)
-                                        {
-                                            script += template
-                                                        .Replace("SCHEMA_NAME", reader["SCHEMA_NAME"])
-                                                        .Replace("COLLECTION_NAME", reader["COLLECTION_NAME"])
-                                                        .Replace("''", "'" + reader["SCRIPT"] + "'");
-
-                                            currentCount++;
-
-                                            helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
-                                        }
+                                        totalCount++;
                                     }
                                 }
                             }
                         }
-                    
-                        return Tuple.Create(script, currentCount);
-                    },
-                    ref counter, ref total, groupToken
-                );
-            }
-        }
 
-        public void DropAll(long token)
-        {
-            helpers.DropObject(
-                token, 
-                ",xsc,", 
-                settings.ScriptPath + settings.XmlSchemaCollectionsFilename, 
-                "XML schema collection", 
-                "drops" + Path.DirectorySeparatorChar.ToString() + "drop-xml-schema-collections.sql"
-                );
-        }
-    
-        public void Restore(long token)
-        {
-            helpers.RestoreObject(
-                token, 
-                ",xsc,", 
-                "xml schema collection", 
-                settings.ScriptPath + settings.XmlSchemaCollectionsFilename
-                );
+                        if (settings.Abort == false)
+                        {
+                            if (reader.Rewind())
+                            {
+                                if (reader.HasRows)
+                                {
+                                    while(reader.Read() && settings.Abort == false)
+                                    {
+                                        script += template
+                                            .Replace("SCHEMA_NAME", reader["SCHEMA_NAME"])
+                                            .Replace("COLLECTION_NAME", reader["COLLECTION_NAME"])
+                                            .Replace("''", "'" + reader["SCRIPT"] + "'");
+
+                                        currentCount++;
+
+                                        helpers.ShowPercentageComplete(token, currentCount, totalCount, startDate, ref lastTimeUpdate, prefix + " ");
+                                    }
+                                }
+                            }
+                        }
+                    }
+                    
+                    return Tuple.Create(script, currentCount);
+                },
+                ref counter, ref total, groupToken
+            );
         }
     }
-}
 
+    public void DropAll(long token)
+    {
+        helpers.DropObject(
+            token, 
+            ",xsc,", 
+            settings.ScriptPath + settings.XmlSchemaCollectionsFilename, 
+            "XML schema collection", 
+            "drops" + Path.DirectorySeparatorChar + "drop-xml-schema-collections.sql"
+        );
+    }
+    
+    public void Restore(long token)
+    {
+        helpers.RestoreObject(
+            token, 
+            ",xsc,", 
+            "xml schema collection", 
+            settings.ScriptPath + settings.XmlSchemaCollectionsFilename
+        );
+    }
+}
