@@ -9,596 +9,577 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Runtime.Versioning;
 using System.Text.RegularExpressions;
-using Fynydd.Halide.Constants;
 
-namespace Fynydd.Halide
+namespace SQribe.Halide.Core;
+
+public static class Identify
 {
-	public static class Identify
+	/// <summary>
+	/// Get OS platform.
+	/// </summary> 
+	/// <returns>OSPlatform object</returns> 
+	public static OSPlatform GetOSPlatform()
 	{
-        /// <summary>
-        /// Get OS platform.
-        /// </summary> 
-        /// <returns>OSPlatform object</returns> 
-        public static OSPlatform GetOSPlatform() 
-        { 
-            OSPlatform osPlatform = OSPlatform.Create("Other platform");
+		OSPlatform osPlatform = OSPlatform.Create("Other platform");
 
-            // Check if it's windows 
-            bool isWindows = RuntimeInformation.IsOSPlatform(OSPlatform.Windows); 
-            osPlatform = isWindows ? OSPlatform.Windows : osPlatform; 
+		// Check if it's windows 
+		bool isWindows = RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
+		osPlatform = isWindows ? OSPlatform.Windows : osPlatform;
 
-            // Check if it's osx 
-            bool isOSX = RuntimeInformation.IsOSPlatform(OSPlatform.OSX); 
-            osPlatform = isOSX ? OSPlatform.OSX : osPlatform; 
+		// Check if it's osx 
+		bool isOSX = RuntimeInformation.IsOSPlatform(OSPlatform.OSX);
+		osPlatform = isOSX ? OSPlatform.OSX : osPlatform;
 
-            // Check if it's Linux 
-            bool isLinux = RuntimeInformation.IsOSPlatform(OSPlatform.Linux); 
-            osPlatform = isLinux ? OSPlatform.Linux : osPlatform;
+		// Check if it's Linux 
+		bool isLinux = RuntimeInformation.IsOSPlatform(OSPlatform.Linux);
+		osPlatform = isLinux ? OSPlatform.Linux : osPlatform;
 
-			return osPlatform; 
-        } 
+		return osPlatform;
+	}
 
-        /// <summary>
-        /// Get OS platform name for output to users.
-        /// </summary> 
-        /// <returns>OSPlatform name (friendly for output)</returns> 
-        public static string GetOSPlatformName()
-        {
-			var result = "Windows";
+	/// <summary>
+	/// Get OS platform name for output to users.
+	/// </summary> 
+	/// <returns>OSPlatform name (friendly for output)</returns> 
+	public static string GetOSPlatformName()
+	{
+		var result = "Windows";
 
-			if (GetOSPlatform() == OSPlatform.OSX)
+		if (GetOSPlatform() == OSPlatform.OSX)
+		{
+			result = "macOS";
+		}
+
+		else if (GetOSPlatform() == OSPlatform.Linux)
+		{
+			result = "Linux";
+		}
+
+		return result;
+	}
+
+	/// <summary>
+	/// Get platform architecture (e.g. x64).
+	/// </summary> 
+	/// <returns>OSPlatform object</returns> 
+	public static string GetPlatformArchitecture()
+	{
+		return RuntimeInformation.ProcessArchitecture.ToString();
+	}
+
+	/// <summary>
+	/// Get the .NET Core runtime version (e.g. "2.2").
+	/// </summary> 
+	/// <returns>String with the .NET Core version number</returns> 
+	public static string GetFrameworkVersion()
+	{
+		var result = Assembly
+			.GetEntryAssembly()?
+			.GetCustomAttribute<TargetFrameworkAttribute>()?
+			.FrameworkName;
+
+		if (string.IsNullOrEmpty(result) == false)
+		{
+			if (result.Contains("Version="))
 			{
-				result = "macOS";
+				result = result.Right("Version=")
+					.TrimStart(new [] {'v'});
+			}
+		}
+
+		return result ?? string.Empty;
+	}
+
+	/// <summary>
+	/// Get the .NET CLR runtime version (e.g. "4.6.27110.04").
+	/// Only works in 2.2 or later.
+	/// </summary> 
+	/// <returns>String with the .NET CLR runtime version number</returns> 
+	public static string GetRuntimeVersion()
+	{
+		return RuntimeInformation.FrameworkDescription.Right(" ");
+	}
+
+	/// <summary>
+	/// Determine if a given string matches a 2-letter state abbreviation.
+	/// </summary>
+	/// <param name="value">2-letter state abbreviation to check.</param>
+	/// <returns>true or false.</returns>
+	public static bool IsState(this string value)
+	{
+		return Geography.StatesAbbreviations.Contains(value.ToUpper());
+	}
+
+	/// <summary>
+	/// Determines if a string value is numeric, but not currency.
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <returns>true or false.</returns>
+	public static bool IsNumeric(this string value)
+	{
+		return double.TryParse(value,
+			(NumberStyles.AllowExponent | NumberStyles.AllowThousands | NumberStyles.AllowDecimalPoint |
+			 NumberStyles.AllowLeadingSign), null, out _);
+	}
+
+	/// <summary>
+	/// Determines if a string value is a bool value or not.
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <returns>true or false.</returns>
+	public static bool IsBool(this string value)
+	{
+		return bool.TryParse(value, out _);
+	}
+
+	/// <summary>
+	/// Determines if a string value can be used as currency.
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <returns>true or false.</returns>
+	public static bool IsCurrency(this string value)
+	{
+		return double.TryParse(value,
+			(NumberStyles.AllowCurrencySymbol | NumberStyles.AllowThousands | NumberStyles.AllowDecimalPoint |
+			 NumberStyles.AllowLeadingSign), CultureInfo.CreateSpecificCulture("en-US"), out _);
+	}
+
+	/// <summary>
+	/// Determines if a string value is numeric, with no symbols, commas, or decimal points.
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <returns>true or false.</returns>
+	public static bool IsPureNumeric(this string value)
+	{
+		return double.TryParse(value, 0, null, out _);
+	}
+
+	/// <summary>
+	/// Checks to see if the passed input has the passed pattern
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <param name="pattern">Regex pattern to use</param>
+	/// <returns>True if the value has the pattern, false otherwise.</returns>
+	public static bool MatchesPattern(this string value, string pattern)
+	{
+		var regEx = new Regex(pattern);
+		return regEx.IsMatch(value);
+	}
+
+	/// <summary>
+	/// Checks the passed input to make sure it validates against all the patterns.
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <param name="patterns">String array of Regex patterns to use</param>
+	/// <returns>True if the input has all of the patterns, false otherwise.</returns>
+	public static bool MatchesPatterns(this string value, string[] patterns)
+	{
+		var result = true;
+
+		foreach (var t in patterns)
+		{
+			if (value.MatchesPattern(t) == false)
+			{
+				result = false;
+			}
+		}
+
+		return result;
+	}
+
+	/// <summary>
+	/// Determines if a string value is a valid date.
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <returns>true or false</returns>
+	public static bool IsDate(this string value)
+	{
+		var result = false;
+
+		if (!string.IsNullOrEmpty(value))
+		{
+			try
+			{
+				// ReSharper disable once UnusedVariable
+				var dt = Convert.ToDateTime(value);
+				result = true;
 			}
 
-			else if (GetOSPlatform() == OSPlatform.Linux)
+			catch
 			{
-				result = "Linux";
+				result = false;
 			}
-
-            return result;
-        }
-
-		/// <summary>
-		/// Get platform architecture (e.g. x64).
-		/// </summary> 
-		/// <returns>OSPlatform object</returns> 
-		public static string GetPlatformArchitecture()
-		{
-			return RuntimeInformation.ProcessArchitecture.ToString();
 		}
 
-		/// <summary>
-		/// Get the .NET Core runtime version (e.g. "2.2").
-		/// </summary> 
-		/// <returns>String with the .NET Core version number</returns> 
-		public static string GetFrameworkVersion()
-        {
-            var result = Assembly
-                .GetEntryAssembly()?
-                .GetCustomAttribute<TargetFrameworkAttribute>()?
-                .FrameworkName;
+		return result;
+	}
 
-            if (string.IsNullOrEmpty(result) == false)
-            {
-                if (result.Contains("Version="))
-                {
-                    result = result.Right("Version=")
-                        .TrimStart(new char[] { 'v' });
-                }
-            }
+	/// <summary>
+	/// Uses RegEx to check for password formatting. Alpha-numeric
+	/// characters and basic typewriter symbols are allowed.
+	/// </summary>
+	/// <param name="value">password string to validate.</param>
+	/// <param name="minLen">Minimum length of valid password; at least 4.</param>
+	/// <param name="maxLen">Maximum length for valid password</param>
+	/// <returns>true if a valid password, false if not.</returns>
+	public static bool IsValidPasswordFormat(this string value, int minLength, int maxLength)
+	{
+		var result = false;
 
-            return result;
-        }
-
-        /// <summary>
-        /// Get the .NET CLR runtime version (e.g. "4.6.27110.04").
-        /// Only works in 2.2 or later.
-        /// </summary> 
-        /// <returns>String with the .NET CLR runtime version number</returns> 
-        public static string GetRuntimeVersion()
-        {
-            return System.Runtime.InteropServices.RuntimeInformation.FrameworkDescription.Right(" ");
-        }
-
-		/// <summary>
-		/// Determine if a given string matches a 2-letter state abbreviation.
-		/// </summary>
-		/// <param name="value">2-letter state abbreviation to check.</param>
-		/// <returns>true or false.</returns>
-		public static bool IsState(this string value)
+		if (string.IsNullOrEmpty(value) == false && minLength > 3 && minLength <= maxLength)
 		{
-			return Geography.StatesAbbreviations.Contains(value.ToUpper());
-		}
-
-		/// <summary>
-		/// Determines if a string value is numeric, but not currency.
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <returns>true or false.</returns>
-		public static bool IsNumeric(this string value)
-		{
-			Double doubleVal;
-			return Double.TryParse(value, (NumberStyles.AllowExponent | NumberStyles.AllowThousands | NumberStyles.AllowDecimalPoint | NumberStyles.AllowLeadingSign), null, out doubleVal);
-		}
-
-		/// <summary>
-		/// Determines if a string value is a bool value or not.
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <returns>true or false.</returns>
-		public static bool IsBool(this string value)
-		{
-			bool boolVal;
-			return bool.TryParse(value, out boolVal);
-		}
-
-		/// <summary>
-		/// Determines if a string value can be used as currency.
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <returns>true or false.</returns>
-		public static bool IsCurrency(this string value)
-		{
-			double doubleVal;
-			return double.TryParse(value, (NumberStyles.AllowCurrencySymbol | NumberStyles.AllowThousands | NumberStyles.AllowDecimalPoint | NumberStyles.AllowLeadingSign), CultureInfo.CreateSpecificCulture("en-US"), out doubleVal);
-		}
-
-		/// <summary>
-		/// Determines if a string value is numeric, with no symbols, commas, or decimal points.
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <returns>true or false.</returns>
-		public static bool IsPureNumeric(this string value)
-		{
-			Double doubleVal;
-			return Double.TryParse(value, 0, null, out doubleVal);
-		}
-
-		/// <summary>
-		/// Checks to see if the passed input has the passed pattern
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <param name="pattern">Regex pattern to use</param>
-		/// <returns>True if the value has the pattern, false otherwise.</returns>
-		public static bool MatchesPattern(this string value, string pattern)
-		{
-			Regex regEx = new Regex(pattern);
-			return regEx.IsMatch(value);
-		}
-
-		/// <summary>
-		/// Checks the passed input to make sure it validates against all the patterns.
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <param name="patterns">String array of Regex patterns to use</param>
-		/// <returns>True if the input has all of the patterns, false otherwise.</returns>
-		public static bool MatchesPatterns(this string value, string[] patterns)
-		{
-			bool result = true;
-
-			for (int i = 0; i < patterns.Length; i++)
+			if (value.Length >= minLength && value.Length <= maxLength)
 			{
-				if (value.MatchesPattern(patterns[i]) == false)
+				if (value.MatchesPattern(RegularExpressions.Password))
 				{
-					result = false;
-				}
-			}
-
-			return result;
-		}
-
-		/// <summary>
-		/// Determines if a string value is a valid date.
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <returns>true or false</returns>
-		public static bool IsDate(this string value)
-		{
-			bool result = false;
-
-			if (!String.IsNullOrEmpty(value))
-			{
-				try
-				{
-					DateTime dt = Convert.ToDateTime(value);
 					result = true;
 				}
+			}
+		}
 
-				catch
+		return result;
+	}
+
+	/// <summary>
+	/// Uses RegEx to check for password formatting. Alpha-numeric
+	/// characters and basic typewriter symbols are allowed.
+	/// Password must be between 8 and 64 characters in length.
+	/// </summary>
+	/// <param name="value">password string to validate.</param>
+	/// <returns>true if a valid password, false if not.</returns>
+	public static bool IsValidPasswordFormat(this string value)
+	{
+		return value.MatchesPattern(RegularExpressions.PasswordAndLength);
+	}
+
+	/// <summary>
+	/// Validate the format of an email address
+	/// </summary>
+	/// <param name="value">email address to evaluate</param>
+	/// <returns>true if formatted as a valid email address, false if not</returns>
+	public static bool IsEmail(this string value)
+	{
+		var result = false;
+
+		if (string.IsNullOrEmpty(value) == false)
+		{
+			result = value.MatchesPattern(RegularExpressions.Email);
+		}
+
+		return result;
+	}
+
+	private class StringComparer : IEqualityComparer<string>
+	{
+		private bool _caseSensitive { get; }
+
+		public StringComparer(bool caseSensitive)
+		{
+			_caseSensitive = caseSensitive;
+		}
+
+		public bool Equals(string? x, string? y)
+		{
+			return _caseSensitive ? x == y : x?.ToLower() == y?.ToLower();
+		}
+
+		public int GetHashCode(string obj)
+		{
+			return int.Parse(obj).ToString().GetHashCode();
+		}
+	}
+
+	/// <summary>
+	/// Verify that a given value is equal to a value in a delimitted list of provided values.
+	/// </summary>
+	/// <param name="value">String to evaluate</param>
+	/// <param name="prevalueList">Delimitted list of possible matches for comparison.</param>
+	/// <param name="delimitter">Delimitter string which defines the boundary between items in the list (e.g. ",")</param>
+	/// <param name="caseSensitive">Set to true to compare values using case sensitivity, or false to ignore case.</param>
+	/// <returns>Empty string on success, or an error message.</returns>
+	public static bool ContainsPrevalue(this string value, string prevalueList, string delimitter,
+		bool caseSensitive = false)
+	{
+		var result = false;
+
+		if (string.IsNullOrEmpty(prevalueList) == false && string.IsNullOrEmpty(value) == false)
+		{
+			var values = prevalueList.Split(new [] {delimitter}, StringSplitOptions.RemoveEmptyEntries);
+
+			if (values.Length > 0)
+			{
+				result = values.Contains(value, new StringComparer(caseSensitive));
+			}
+		}
+
+		return result;
+	}
+
+	/// <summary>
+	/// Run data type and field length validation on a specific value.
+	/// Certain types of validators obviate the need to specify a minimum or maximum length,
+	/// like ValidationOptions.Email.
+	/// </summary>
+	/// <param name="value">Value to validate.</param>
+	/// <param name="valType">Constant determining what type of validation.</param>
+	/// <param name="minLength">Minimum length alowed.</param>
+	/// <param name="maxLength">Maximum length allowed.</param>
+	/// <param name="optional">Field is optional. Zero length validates, otherwise, full validation occurs.</param>
+	/// <returns>Empty string if validation succeeds, error message on failure.</returns>
+	public static string ValidateText(this string value, ValidationOptions valType, int minLength = 0, int maxLength = 0, bool optional = false)
+	{
+		var result = string.Empty;
+
+		if (optional && value.Length < 1) return result;
+
+		// Evaluate length of value if minLength and maxLengths are non-zero
+		if (minLength > 0 && maxLength > 0)
+		{
+			if (value.Length > maxLength)
+			{
+				result += "Must be " + maxLength.ToString() + " character(s) long";
+			}
+			else
+			{
+				if (value.Length < minLength)
 				{
-					result = false;
+					result += "Must be " + minLength.ToString() + " character(s) or longer";
 				}
 			}
-
-			return result;
 		}
 
-		/// <summary>
-		/// Uses RegEx to check for password formatting. Alpha-numeric
-		/// characters and basic typewriter symbols are allowed.
-		/// </summary>
-		/// <param name="value">password string to validate.</param>
-		/// <param name="minLen">Minimum length of valid password; at least 4.</param>
-		/// <param name="maxLen">Maximum length for valid password</param>
-		/// <returns>true if a valid password, false if not.</returns>
-		public static bool IsValidPasswordFormat(this string value, int minLength, int maxLength)
+		// If no length errors, validate data is of a specific format
+		if (result == string.Empty)
 		{
-			bool result = false;
-
-			if (string.IsNullOrEmpty(value) == false && minLength > 3 && minLength <= maxLength)
+			switch (valType)
 			{
-				if (value.Length >= minLength && value.Length <= maxLength)
+				case ValidationOptions.USStateAbbreviation:
 				{
-					if (value.MatchesPattern(RegularExpressions.Password))
-					{
-						result = true;
-					}
-				}
-			}
+					var flag = Geography.StatesAbbreviations.Contains(value, new StringComparer(false));
 
-			return result;
-		}
-
-		/// <summary>
-		/// Uses RegEx to check for password formatting. Alpha-numeric
-		/// characters and basic typewriter symbols are allowed.
-		/// Password must be between 8 and 64 characters in length.
-		/// </summary>
-		/// <param name="value">password string to validate.</param>
-		/// <returns>true if a valid password, false if not.</returns>
-		public static bool IsValidPasswordFormat(this string value)
-		{
-			return value.MatchesPattern(RegularExpressions.PasswordAndLength);
-		}
-
-		/// <summary>
-		/// Validate the format of an email address
-		/// </summary>
-		/// <param name="value">email address to evaluate</param>
-		/// <returns>true if formatted as a valid email address, false if not</returns>
-		public static bool IsEmail(this string value)
-		{
-			bool result = false;
-
-			if (string.IsNullOrEmpty(value) == false)
-			{
-				result = value.MatchesPattern(RegularExpressions.Email);
-			}
-
-			return result;
-		}
-
-		private class StringComparer : IEqualityComparer<string>
-		{
-			private bool _caseSensitive { get; set; }
-
-			public StringComparer(bool caseSensitive)
-			{
-				_caseSensitive = caseSensitive;
-			}
-
-			public StringComparer()
-			{
-				_caseSensitive = false;
-			}
-
-			public bool Equals(string x, string y)
-			{
-				return ((_caseSensitive ? x == y : x.ToLower() == y.ToLower()));
-			}
-
-			public int GetHashCode(string obj)
-			{
-				return int.Parse(obj).ToString().GetHashCode();
-			}
-		}
-
-		/// <summary>
-		/// Verify that a given value is equal to a value in a delimitted list of provided values.
-		/// </summary>
-		/// <param name="value">String to evaluate</param>
-		/// <param name="prevalueList">Delimitted list of possible matches for comparison.</param>
-		/// <param name="delimitter">Delimitter string which defines the boundary between items in the list (e.g. ",")</param>
-		/// <param name="caseSensitive">Set to true to compare values using case sensitivity, or false to ignore case.</param>
-		/// <returns>Empty string on success, or an error message.</returns>
-		public static bool ContainsPrevalue(this string value, string prevalueList, string delimitter, bool caseSensitive = false)
-		{
-			bool result = false;
-
-			if (string.IsNullOrEmpty(prevalueList) == false && string.IsNullOrEmpty(value) == false)
-			{
-				string[] values = prevalueList.Split(new string[] { delimitter }, StringSplitOptions.RemoveEmptyEntries);
-
-				if (values.Length > 0)
-				{
-					result = values.Contains(value, new StringComparer(caseSensitive));
-				}
-			}
-
-			return result;
-		}
-
-		/// <summary>
-		/// Run data type and field length validation on a specific value.
-		/// Certain types of validators obviate the need to specify a minimum or maximum length,
-		/// like ValidationOptions.Email.
-		/// </summary>
-		/// <param name="value">Value to validate.</param>
-		/// <param name="valType">Constant determining what type of validation.</param>
-		/// <param name="minLength">Minimum length alowed.</param>
-		/// <param name="maxLength">Maximum length allowed.</param>
-		/// <param name="optional">Field is optional. Zero length validates, otherwise, full validation occurs.</param>
-		/// <returns>Empty string if validation succeeds, error message on failure.</returns>
-		public static string ValidateText(this string value, ValidationOptions valType, int minLength, int maxLength, bool optional)
-		{
-			string result = "";
-
-			if (optional && value.Length < 1) return result;
-
-			// Evaluate length of value if minLength and maxLengths are non-zero
-			if (minLength > 0 && maxLength > 0)
-			{
-				if (value.Length > maxLength)
-				{
-					result += "Must be " + maxLength.ToString() + " character(s) long";
-				}
-				else
-				{
-					if (value.Length < minLength)
-					{
-						result += "Must be " + minLength.ToString() + " character(s) or longer";
-					}
-				}
-			}
-
-			// If no length errors, validate data is of a specific format
-			if (result == "")
-			{
-				if (valType == ValidationOptions.USStateAbbreviation)
-				{
-					bool flag = Geography.StatesAbbreviations.Contains(value, new StringComparer(false));
-
-					if (!flag)
+					if (flag == false)
 					{
 						result += "Not a state";
 					}
-				}
 
-				if (valType == ValidationOptions.Email)
+					break;
+				}
+				case ValidationOptions.Email:
 				{
-					if (!IsEmail(value))
+					if (IsEmail(value) == false)
 					{
 						result += "Invalid email address format";
 					}
-				}
 
-				if (valType == ValidationOptions.EmailList)
+					break;
+				}
+				// Invalid string
+				case ValidationOptions.EmailList when value.Length < 6:
+					result += "Invalid email address format";
+					break;
+				case ValidationOptions.EmailList:
 				{
-					// Invalid string
-					if (value.Length < 6)
+					// Only one email address? If not, check for multiple.
+					if (IsEmail(value) == false)
 					{
-						result += "Invalid email address format";
-					}
-					else
-					{
-						// Only one email address? If not, check for multiple.
-						if (!IsEmail(value))
+						var delim = string.Empty;
+
+						if (delim == string.Empty && value.IndexOf(",", StringComparison.Ordinal) > 0) delim = ",";
+						if (delim == string.Empty && value.IndexOf("|", StringComparison.Ordinal) > 0) delim = "|";
+						if (delim == string.Empty && value.IndexOf(";", StringComparison.Ordinal) > 0) delim = ";";
+						if (delim == string.Empty && value.IndexOf("/", StringComparison.Ordinal) > 0) delim = "/";
+						if (delim == string.Empty && value.IndexOf(" ", StringComparison.Ordinal) > 0) delim = " ";
+
+						if (delim != string.Empty)
 						{
-							string delim = "";
+							value = value.Replace(delim, ",");
+							value = value.Replace(" ", string.Empty);
 
-							if (delim == "" && value.IndexOf(",") > 0) delim = ",";
-							if (delim == "" && value.IndexOf("|") > 0) delim = "|";
-							if (delim == "" && value.IndexOf(";") > 0) delim = ";";
-							if (delim == "" && value.IndexOf("/") > 0) delim = "/";
-							if (delim == "" && value.IndexOf(" ") > 0) delim = " ";
+							while (value.IndexOf(",,", StringComparison.Ordinal) > 0) value = value.Replace(",,", ",");
 
-							if (delim != "")
+							var addresses = value.Split(new[] {','});
+
+							foreach (var t in addresses)
 							{
-								value = value.Replace(delim, ",");
-								value = value.Replace(" ", "");
-
-								while (value.IndexOf(",,") > 0) value = value.Replace(",,", ",");
-
-								string[] addresses = value.Split(new[] { ',' });
-
-								for (int x = 0; x < addresses.Length; x++)
+								if (!IsEmail(t))
 								{
-									if (!IsEmail(addresses[x]))
-									{
-										result += "One or more email addresses are invalid";
-									}
+									result += "One or more email addresses are invalid";
 								}
 							}
-							else
-							{
-								result += "Must have one or more email addresses";
-							}
+						}
+						else
+						{
+							result += "Must have one or more email addresses";
 						}
 					}
-				}
 
-				if (valType == ValidationOptions.Number)
+					break;
+				}
+				case ValidationOptions.Number:
 				{
-					if (!IsNumeric(value))
+					if (IsNumeric(value) == false)
 					{
 						result += "Not a numeric value";
 					}
-				}
 
-				if (valType == ValidationOptions.PureNumber)
+					break;
+				}
+				case ValidationOptions.PureNumber:
 				{
-					if (!IsPureNumeric(value))
+					if (IsPureNumeric(value) == false)
 					{
 						result += "Must only contain numbers";
 					}
-				}
 
-				if (valType == ValidationOptions.NonZero)
-				{
-					if (!IsPureNumeric(value))
-					{
-						result += "Must only contain numbers";
-					}
-					else
-					{
-						if (Convert.ToInt32(value) <= 0)
-						{
-							result += "Must be greater than zero";
-						}
-					}
+					break;
 				}
-
-				if (valType == ValidationOptions.Currency)
+				case ValidationOptions.NonZero when IsPureNumeric(value) == false:
+					result += "Must only contain numbers";
+					break;
+				case ValidationOptions.NonZero:
 				{
-					if (!IsCurrency(value))
+					if (Convert.ToInt32(value) <= 0)
+					{
+						result += "Must be greater than zero";
+					}
+
+					break;
+				}
+				case ValidationOptions.Currency:
+				{
+					if (IsCurrency(value) == false)
 					{
 						result += "Not a currency value";
 					}
-				}
 
-				if (valType == ValidationOptions.Date)
+					break;
+				}
+				case ValidationOptions.Date:
 				{
-					if (!IsDate(value))
+					if (IsDate(value) == false)
 					{
 						result += "Not a date";
 					}
-				}
 
-				if (valType == ValidationOptions.Telephone)
+					break;
+				}
+				case ValidationOptions.Telephone:
 				{
-					if (!MatchesPattern(value, RegularExpressions.TelephoneUS))
+					if (MatchesPattern(value, RegularExpressions.TelephoneUS) == false)
 					{
 						result += "Not a telephone number";
 					}
-				}
 
-				if (valType == ValidationOptions.Telephone10)
+					break;
+				}
+				case ValidationOptions.Telephone10:
 				{
-					if (!MatchesPattern(value, RegularExpressions.Telephone10))
+					if (MatchesPattern(value, RegularExpressions.Telephone10) == false)
 					{
 						result += "Not a telephone number with area code";
 					}
-				}
 
-				if (valType == ValidationOptions.CreditCard)
+					break;
+				}
+				case ValidationOptions.CreditCard:
 				{
-					if (!MatchesPattern(value, RegularExpressions.ValidCreditCardNumber))
+					if (MatchesPattern(value, RegularExpressions.ValidCreditCardNumber) == false)
 					{
 						result += "Not a credit card number";
 					}
-				}
 
-				if (valType == ValidationOptions.ZipCode)
+					break;
+				}
+				case ValidationOptions.ZipCode:
 				{
-					if (!MatchesPattern(value, RegularExpressions.ZipCode))
+					if (MatchesPattern(value, RegularExpressions.ZipCode) == false)
 					{
 						result += "Not a 5-digit zip code";
 					}
-				}
 
-				if (valType == ValidationOptions.Url)
+					break;
+				}
+				case ValidationOptions.Url:
 				{
-					if (!MatchesPattern(value, RegularExpressions.Url))
+					if (MatchesPattern(value, RegularExpressions.Url) == false)
 					{
 						result += "Not a URL";
 					}
-				}
 
-				if (valType == ValidationOptions.IPv4Address)
+					break;
+				}
+				case ValidationOptions.IPv4Address:
 				{
-					if (!MatchesPattern(value, RegularExpressions.IPv4Address))
+					if (MatchesPattern(value, RegularExpressions.IPv4Address) == false)
 					{
 						result += "Not an IP address";
 					}
-				}
 
-				if (valType == ValidationOptions.Ssn)
+					break;
+				}
+				case ValidationOptions.Ssn:
 				{
-					if (!MatchesPattern(value, RegularExpressions.SocialSecurityNumber))
+					if (MatchesPattern(value, RegularExpressions.SocialSecurityNumber) == false)
 					{
 						result += "Not a social security number";
 					}
-				}
 
-				if (valType == ValidationOptions.Time)
+					break;
+				}
+				case ValidationOptions.Time:
 				{
-					if (!MatchesPattern(value, RegularExpressions.Time) && !MatchesPattern(value, RegularExpressions.TimeMilitary))
+					if (MatchesPattern(value, RegularExpressions.Time) == false && MatchesPattern(value, RegularExpressions.TimeMilitary) == false)
 					{
 						result += "Not a time of day";
 					}
-				}
 
-				if (valType == ValidationOptions.Domain)
+					break;
+				}
+				case ValidationOptions.Domain:
 				{
-					if (!MatchesPattern(value, RegularExpressions.DomainName))
+					if (MatchesPattern(value, RegularExpressions.DomainName) == false)
 					{
 						result += "Not a domain name";
 					}
-				}
 
-				if (valType == ValidationOptions.FullNameWithComma)
+					break;
+				}
+				case ValidationOptions.FullNameWithComma:
 				{
-					if (!MatchesPattern(value, RegularExpressions.FullNameWithComma))
+					if (MatchesPattern(value, RegularExpressions.FullNameWithComma) == false)
 					{
 						result += "Not last name, first name";
 					}
+
+					break;
 				}
 			}
-
-			return result;
 		}
 
-		/// <summary>
-		/// Run data type and field length validation on a specific value.
-		/// Certain types of validators obviate the need to specify a minimum or maximum length,
-		/// like ValidationOptions.Email.
-		/// </summary>
-		/// <param name="value">Value to validate.</param>
-		/// <param name="valType">Constant determining what type of validation.</param>
-		/// <param name="minLength">Minimum length alowed.</param>
-		/// <param name="maxLength">Maximum length allowed.</param>
-		/// <returns>Empty string if validation succeeds, error message on failure.</returns>
-		public static string ValidateText(this string value, ValidationOptions valType, int minLength, int maxLength)
-		{
-			return ValidateText(value, valType, minLength, maxLength, false);
-		}
+		return result;
+	}
 
-		/// <summary>
-		/// Run data type and field length validation on a specific value.
-		/// Certain types of validators obviate the need to specify a minimum or maximum length,
-		/// like ValidationOptions.Email.
-		/// </summary>
-		/// <param name="value">Value to validate.</param>
-		/// <param name="valType">Constant determining what type of validation.</param>
-		/// <returns>Empty string if validation succeeds, error message on failure.</returns>
-		public static string ValidateText(this string value, ValidationOptions valType)
-		{
-			return ValidateText(value, valType, 0, 0, false);
-		}
+	/// <summary>
+	/// Run field length validation on a specific value.
+	/// </summary>
+	/// <param name="value">Value to validate.</param>
+	/// <param name="minLength">Minimum length alowed.</param>
+	/// <param name="maxLength">Maximum length allowed.</param>
+	/// <returns>Empty string if validation succeeds, error message on failure.</returns>
+	public static string ValidateText(this string value, int minLength, int maxLength)
+	{
+		return ValidateText(value, ValidationOptions.Length, minLength, maxLength);
+	}
 
-		/// <summary>
-		/// Run field length validation on a specific value.
-		/// </summary>
-		/// <param name="value">Value to validate.</param>
-		/// <param name="minLength">Minimum length alowed.</param>
-		/// <param name="maxLength">Maximum length allowed.</param>
-		/// <returns>Empty string if validation succeeds, error message on failure.</returns>
-		public static string ValidateText(this string value, int minLength, int maxLength)
-		{
-			return ValidateText(value, ValidationOptions.Length, minLength, maxLength, false);
-		}
-
-		/// <summary>
-		/// Run data type validation on a specific value.
-		/// </summary>
-		/// <param name="value">Value to validate.</param>
-		/// <param name="valType">Constant determining what type of validation.</param>
-		/// <param name="optional">Field is optional. Zero length validates, otherwise, full validation occurs.</param>
-		/// <returns>Empty string if validation succeeds, error message on failure.</returns>
-		public static string ValidateText(this string value, ValidationOptions valType, bool optional)
-		{
-			return ValidateText(value, valType, 0, 0, optional);
-		}
+	/// <summary>
+	/// Run data type validation on a specific value.
+	/// </summary>
+	/// <param name="value">Value to validate.</param>
+	/// <param name="valType">Constant determining what type of validation.</param>
+	/// <param name="optional">Field is optional. Zero length validates, otherwise, full validation occurs.</param>
+	/// <returns>Empty string if validation succeeds, error message on failure.</returns>
+	public static string ValidateText(this string value, ValidationOptions valType, bool optional)
+	{
+		return ValidateText(value, valType, 0, 0, optional);
 	}
 }
